+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 ### unreleased
+- **fix:** refactor routeMap -> routeList
 
 ### [v0.3.1] - 2016-05-17
 - **doc:** add explanation for options.log